+++ tidy3d/components/medium.py
@@ -381,12 +381,13 @@ class PoleResidue(DispersiveMedium):
     """A dispersive medium described by the pole-residue pair model.
     The frequency-dependence of the complex-valued permittivity is described by:
 
+    Note
+    ----
     .. math::
+
         \\epsilon(\\omega) = \\epsilon_\\infty - \\sum_i
         \\left[\\frac{c_i}{j \\omega + a_i} +
         \\frac{c_i^*}{j \\omega + a_i^*}\\right]
-
-    where :math:`a_i` and :math:`c_i` are in units of rad/s.
 
     Example
     -------
@@ -406,19 +407,6 @@ class PoleResidue(DispersiveMedium):
         description="List of complex-valued (:math:`a_i, c_i`) poles for the model.",
         units=RADPERSEC,
     )
-
-    # @pydantic.validator("poles", always=True)
-    # def convert_complex(cls, val):
-    #     """convert list of poles to complex"""
-
-    #     poles_complex = []
-    #     for (a, c) in val:
-    #         if isinstance(a, ComplexNumber):
-    #             a = a.real + 1j * a.imag
-    #         if isinstance(c, ComplexNumber):
-    #             c = c.real + 1j * c.imag
-    #         poles_complex.append((a, c))
-    #     return poles_complex
 
     @ensure_freq_in_range
     def eps_model(self, frequency: float) -> complex:
@@ -462,11 +450,11 @@ class Sellmeier(DispersiveMedium):
     """A dispersive medium described by the Sellmeier model.
     The frequency-dependence of the refractive index is described by:
 
+    Note
+    ----
     .. math::
 
         n(\\lambda)^2 = 1 + \\sum_i \\frac{B_i \\lambda^2}{\\lambda^2 - C_i}
-
-    where :math:`\\lambda` is in microns, :math:`B_i` is unitless and :math:`C_i` is in microns^2.
 
     Example
     -------
@@ -475,7 +463,8 @@ class Sellmeier(DispersiveMedium):
     """
 
     coeffs: List[Tuple[float, PositiveFloat]] = pydantic.Field(
-        title="Coefficients", description="List of Sellmeier (:math:`B_i, C_i`) coefficients."
+        title="Coefficients",
+        description="List of Sellmeier (:math:`B_i, C_i`) coefficients (unitless, microns^2).",
     )
 
     def _n_model(self, frequency: float) -> complex:
@@ -523,11 +512,12 @@ class Lorentz(DispersiveMedium):
     """A dispersive medium described by the Lorentz model.
     The frequency-dependence of the complex-valued permittivity is described by:
 
+    Note
+    ----
     .. math::
+
         \\epsilon(f) = \\epsilon_\\infty + \\sum_i
         \\frac{\\Delta\\epsilon_i f_i^2}{f_i^2 - 2jf\\delta_i - f^2}
-
-    where :math:`f, f_i, \\delta_i` are in Hz.
 
     Example
     -------
@@ -544,7 +534,7 @@ class Lorentz(DispersiveMedium):
     coeffs: List[Tuple[float, float, float]] = pydantic.Field(
         ...,
         title="Epsilon at Infinity",
-        description="List of (:math:`\\Delta\\epsilon_i, f_i, \\delta_i`) values for model.",
+        description="List of (:math:`\\Delta\\epsilon_i, f_i, \\delta_i`) values for model (Hz).",
     )
 
     @ensure_freq_in_range
@@ -572,18 +562,12 @@ class Lorentz(DispersiveMedium):
                 c0 = de * w ** 2 / 4 / r
                 a1 = -d - r
                 c1 = -c0
-                # a0 = self.complex_to_tuple(a0)
-                # c0 = self.complex_to_tuple(c0)
-                # a1 = self.complex_to_tuple(a1)
-                # c1 = self.complex_to_tuple(c1)
                 poles.append((a0, c0))
                 poles.append((a1, c1))
             else:
                 r = np.sqrt(w * w - d * d)
                 a = -d - 1j * r
                 c = 1j * de * w ** 2 / 2 / r
-                # a = self.complex_to_tuple(a)
-                # c = self.complex_to_tuple(c)
                 poles.append((a, c))
 
         return PoleResidue(
@@ -598,11 +582,12 @@ class Drude(DispersiveMedium):
     """A dispersive medium described by the Drude model.
     The frequency-dependence of the complex-valued permittivity is described by:
 
+    Note
+    ----
     .. math::
+
         \\epsilon(f) = \\epsilon_\\infty - \\sum_i
         \\frac{ f_i^2}{f^2 + jf\\delta_i}
-
-    where :math:`f, f_i, \\delta_i` are in Hz.
 
     Example
     -------
@@ -617,7 +602,9 @@ class Drude(DispersiveMedium):
     )
 
     coeffs: List[Tuple[float, PositiveFloat]] = pydantic.Field(
-        ..., title="Coefficients", description="List of (:math:`f_i, \\delta_i`) values for model."
+        ...,
+        title="Coefficients",
+        description="List of (:math:`f_i, \\delta_i`) values for model (Hz).",
     )
 
     @ensure_freq_in_range
@@ -645,11 +632,6 @@ class Drude(DispersiveMedium):
             c1 = -c0
             a1 = -d + 0j
 
-            # a0 = self.complex_to_tuple(a0)
-            # c0 = self.complex_to_tuple(c0)
-            # a1 = self.complex_to_tuple(a1)
-            # c1 = self.complex_to_tuple(c1)
-
             poles.append((a0, c0))
             poles.append((a1, c1))
 
@@ -665,11 +647,12 @@ class Debye(DispersiveMedium):
     """A dispersive medium described by the Debye model.
     The frequency-dependence of the complex-valued permittivity is described by:
 
+    Note
+    ----
     .. math::
+
         \\epsilon(f) = \\epsilon_\\infty + \\sum_i
         \\frac{\\Delta\\epsilon_i}{1 - jf\\tau_i}
-
-    where :math:`f` is in Hz, and :math:`\\tau_i` is in seconds.
 
     Example
     -------
@@ -686,7 +669,7 @@ class Debye(DispersiveMedium):
     coeffs: List[Tuple[float, PositiveFloat]] = pydantic.Field(
         ...,
         title="Coefficients",
-        description="List of (:math:`\\Delta\\epsilon_i, \\tau_i`) values for model.",
+        description="List of (:math:`\\Delta\\epsilon_i, \\tau_i`) values for model (Hz, sec).",
     )
 
     @ensure_freq_in_range