+++ tidy3d/plugins/__init__.py
@@ -1,5 +1,6 @@
 # import the specific classes / functions needed for the plugins
 
 from .dispersion.fit import DispersionFitter
+from .dispersion.fit_web import StableDispersionFitter
 from .mode.mode_solver import ModeSolver
 from .near2far.near2far import Near2Far