+++ tidy3d/web/container.py
@@ -138,6 +138,8 @@ class Job(WebContainer):
 
             while status not in ("success", "error", "diverged", "deleted", "draft"):
                 new_status = self.status
+                if new_status == "visualize":
+                    new_status = "success"
                 if new_status != status:
                     console.log(f"status = {new_status}")
                     status = new_status
@@ -324,6 +326,8 @@ class Batch(WebContainer):
                 for status_index, (task_name, job) in enumerate(self.jobs.items()):
                     current_status = statuses[status_index]
                     new_status = job.status
+                    if new_status == "visualize":
+                        new_status = "success"
                     if new_status != current_status:
                         completed = run_statuses.index(new_status)
                         pbar = pbar_tasks[task_name]
@@ -401,7 +405,7 @@ class Batch(WebContainer):
         current iteration in memory.
         """
         sim_data_dir = {}
-        self.download(path_dir=path_dir)
+        # self.download(path_dir=path_dir)
         for task_name, job in self.jobs.items():
             job_path = self._job_data_path(task_id=job.task_id, path_dir=path_dir)
             sim_data = job.load(path=job_path, normalize_index=normalize_index)